# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zernio Gateway API",
    "version": "0.1.0",
    "x-gateway-contract-version": "1",
    "x-capability-contracts": [
      "social.publishing@1",
      "social.scheduling@1"
    ],
    "description": "Zeptly's gateway to Zernio. Implements Gateway Contract v1 (/v1/gateway, /v1/capabilities, /v1/connections) and the capability contracts Zernio backs in this gateway: Social Publishing v1 (/v1/social/publishing) and Social Scheduling v1. Provider identifiers never appear in responses. Authenticated with ZS1-HMAC-SHA256 service signatures (see docs/SECURITY.md)."
  },
  "components": {
    "securitySchemes": {
      "zeptlyServiceSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Zeptly-Signature",
        "description": "ZS1-HMAC-SHA256 signature. Also requires X-Zeptly-Caller, X-Zeptly-Timestamp and (workspace routes) X-Zeptly-Workspace-Id. See docs/SECURITY.md."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Canonical error code (see GET /v1/gateway → errorCodes and docs/GATEWAY-CONTRACT.md)"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "attempts": {
            "type": "number"
          },
          "maxAttempts": {
            "type": "number"
          },
          "runAt": {
            "type": "string"
          },
          "lastError": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "attempts",
          "maxAttempts",
          "runAt",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "WebhookEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "eventType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "attempts": {
            "type": "number"
          },
          "lastError": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "eventType",
          "status",
          "attempts",
          "receivedAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v1/admin/jobs": {
      "get": {
        "summary": "Inspect jobs (e.g. status=dead for dead letters)",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "running",
                "succeeded",
                "dead"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/jobs/{id}/retry": {
      "post": {
        "summary": "Requeue a dead job",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "requeued": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "requeued"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/webhook-events": {
      "get": {
        "summary": "Inspect received webhooks",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "received",
                "processing",
                "processed",
                "ignored",
                "failed"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WebhookEvent"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/reconcile": {
      "post": {
        "summary": "Trigger reconciliation (all workspaces) after an incident",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enqueued": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "enqueued"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
...
}